requirements.txt
asyncio==3.4.3
chromadb==0.6.3
click==8.1.8
langchain==0.3.17
langchain-community==0.3.16
langchain-experimental==0.3.4
langchain-openai==0.3.5
loguru==0.7.3
networkx==3.4.2
openai==1.61.1
openpyxl==3.1.5
pandas==2.2.3
plotly==6.0.1
python-dotenv==1.0.1
selenium==4.29.0
setuptools==75.8.0
unstructured==0.16.20
zaproxy==0.4.0